Database

+++ ui/src/main/java/net/chmielowski/baggage/ui/EquipmentId.kt
@@ -1,3 +1,3 @@
 package net.chmielowski.baggage.ui
 
-data class EquipmentId(val value: Int)
+data class EquipmentId(val value: Long)
